backend/requirements.txt
aiodns==3.5.0
aiofiles==0.8.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bbox-visualizer==0.1.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
fastapi==0.118.0
fonttools==4.60.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
kiwisolver==1.4.9
landingai==0.3.49
matplotlib==3.10.6
multidict==6.6.4
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.3
pillow==10.4.0
propcache==0.3.2
psutil==7.1.0
pycares==4.11.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pymongo==4.15.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-toolbelt==1.0.0
segmentation-mask-overlay==0.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==8.5.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.37.0
yarl==1.20.1
